Move helper function to its own file so it can be tested
